+++ CHANGELOG.md
@@ -25,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PHP 8.0 support
 - Migration trigger on module activation
 
+### Fixed
+- Add error message to product response when variant loading is disabled [#0007421](https://bugs.oxid-esales.com/view.php?id=7421)
+
 ## [3.0.0] - 2023-06-08
 
 ### Added